Extract the ConfigSetting implementation into its own file.
