+++ Core/Core.vcxproj.filters
@@ -1195,6 +1195,9 @@
     <ClCompile Include="MIPS\MIPSVFPUFallbacks.cpp">
       <Filter>MIPS</Filter>
     </ClCompile>
+    <ClCompile Include="ConfigSettings.cpp">
+      <Filter>Core</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ELF\ElfReader.h">
@@ -1931,6 +1934,9 @@
     </ClInclude>
     <ClInclude Include="MIPS\MIPSVFPUFallbacks.h">
       <Filter>MIPS</Filter>
+    </ClInclude>
+    <ClInclude Include="ConfigSettings.h">
+      <Filter>Core</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
